move UI classes
