+++ src/TileGame/MouseManager.java
@@ -1,5 +1,7 @@
 package TileGame;
 
+import Utils.UIManager;
+
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseListener;
 import java.awt.event.MouseMotionListener;
