+++ src/Utils/UIObject.java
@@ -1,4 +1,4 @@
-package TileGame;
+package Utils;
 
 import java.awt.*;
 import java.awt.event.MouseEvent;
